Enabled streaming i/o in ImageRoutes. Refactored common code to new class DicomStreams
