+++ src/main/scala/se/nimsa/sbx/app/routing/ImageRoutes.scala
@@ -26,15 +26,16 @@ import akka.http.scaladsl.model.{ContentTypes, HttpEntity}
 import akka.http.scaladsl.server.Directives._
 import akka.http.scaladsl.server.Route
 import akka.pattern.ask
-import akka.stream.scaladsl.{Sink, SourceQueueWithComplete, StreamConverters, Source => StreamSource}
+import akka.stream.scaladsl.{Sink, SourceQueueWithComplete, Source => StreamSource}
 import akka.stream.{OverflowStrategy, QueueOfferResult}
 import akka.util.ByteString
 import org.dcm4che3.data.Attributes
-import se.nimsa.sbx.anonymization.AnonymizationProtocol.ReverseAnonymization
+import org.dcm4che3.io.DicomStreamException
 import se.nimsa.sbx.app.GeneralProtocol._
 import se.nimsa.sbx.app.SliceboxBase
 import se.nimsa.sbx.dicom.DicomHierarchy.{FlatSeries, Image, Patient, Study}
 import se.nimsa.sbx.dicom._
+import se.nimsa.sbx.dicom.streams.DicomStreams._
 import se.nimsa.sbx.metadata.MetaDataProtocol._
 import se.nimsa.sbx.storage.StorageProtocol._
 import se.nimsa.sbx.user.UserProtocol.ApiUser
@@ -169,30 +170,22 @@ trait ImageRoutes {
       }
     }
 
-  private def addDicomDataRoute(bytes: StreamSource[ByteString, Any], apiUser: ApiUser) = {
-    val is = bytes.runWith(StreamConverters.asInputStream())
-    val dicomData = DicomUtil.loadDicomData(is, withPixelData = true)
+  private def addDicomDataRoute(bytes: StreamSource[ByteString, Any], apiUser: ApiUser): Route = {
+
     val source = Source(SourceType.USER, apiUser.user, apiUser.id)
-    val futureImageAndOverwrite =
-      storageService.ask(CheckDicomData(dicomData, useExtendedContexts = true)).mapTo[Boolean].flatMap {
-        _ =>
-          anonymizationService.ask(ReverseAnonymization(dicomData.attributes)).mapTo[Attributes].flatMap {
-            reversedAttributes =>
-              metaDataService.ask(AddMetaData(reversedAttributes, source)).mapTo[MetaDataAdded].flatMap {
-                metaData =>
-                  storageService.ask(AddDicomData(dicomData.copy(attributes = reversedAttributes), source, metaData.image)).mapTo[DicomDataAdded].map {
-                    dicomDataAdded =>
-                      (metaData.image, dicomDataAdded.overwrite)
-                  }
-              }
+    val tmpPath = createTempPath()
+    val futureUpload = bytes.runWith(uploadSink(tmpPath, storage, anonymizationService))
+
+    onSuccess(futureUpload) {
+      case (_, dicomData) =>
+        val attributes: Attributes = dicomData._2.getOrElse(throw new DicomStreamException("DICOM data has no dataset"))
+        onSuccess(metaDataService.ask(AddMetaData(attributes, source)).mapTo[MetaDataAdded]) { metaData =>
+          onSuccess(storageService.ask(MoveDicomData(tmpPath, s"${metaData.image.id}")).mapTo[DicomDataMoved]) { _ =>
+            system.eventStream.publish(ImageAdded(metaData.image, source, !metaData.imageAdded))
+            val httpStatus = if (metaData.imageAdded) Created else OK
+            complete((httpStatus, metaData.image))
           }
-      }
-    onSuccess(futureImageAndOverwrite) {
-      case (image, overwrite) =>
-        if (overwrite)
-          complete((OK, image))
-        else
-          complete((Created, image))
+        }
     }
   }
